misc: move sensor decorators to same file
